Refactor ToyConvModel into shared testing architectures
