+++ test/quantization/quantize_/workflows/float8/test_float8_tensor.py
@@ -29,6 +29,7 @@ from torchao.quantization import (
 )
 from torchao.quantization.quantize_.common import KernelPreference
 from torchao.quantization.utils import compute_error
+from torchao.testing.model_architectures import ToyConvModel
 from torchao.testing.utils import TorchAOIntegrationTestCase
 from torchao.utils import (
     _is_mslk_available,
@@ -73,26 +74,6 @@ class ToyLinearModel(torch.nn.Module):
             assert granularity == (PerBlock([1, 128]), PerBlock([128, 128]))
             assert qs1.shape == (N // 128, K // 128)
             assert qs2.shape == (K // 128, N // 128)
-
-
-class ToyConvModel(torch.nn.Module):
-    def __init__(
-        self, dim, in_channels, out_channels, kernel_size, bias, padding, dtype, device
-    ):
-        super().__init__()
-        convs = {1: torch.nn.Conv1d, 2: torch.nn.Conv2d, 3: torch.nn.Conv3d}
-        self.conv = convs[dim](
-            in_channels,
-            out_channels,
-            kernel_size,
-            bias=bias,
-            padding=padding,
-            dtype=dtype,
-            device=device,
-        )
-
-    def forward(self, x):
-        return self.conv(x)
 
 
 class GroupedMMModel(torch.nn.Module):
